Add functions for downloading data
Stengel is heavily reliant on external data sources for model training.
The functions added here automate the process of downloading the data
from third-party websites.

This is also the first non-trivial commit, so it sets up the project
structure. stengel/ is a package defining functions and classes used by
the rest of the project; test/ contains automated unit tests for
the stengel package.