Add class for game metadata
stengel/sim/metadata is the first file in stengel/sim, so a quick word
about the new folder: sim contains classes for representing objects and
events in baseball games, along with the logic to keep track of the
state of the game, given that events are provided. The intent of sim at
this point is not to be able to generate new events, much less simulate
entirely new games de novo.

GameMetadata (and most other classes in stengel/sim) inherit from
DictSerializable, which is so that historical games can be stored in
MongoDB/CouchDB, or any other NoSQL database that represents data as
JSON documents.