Add PitchFx integration logic
Adding data from PitchFx to a Game object is non-trivial, since the
pitches from the game are from a different source -- there is a
significant amount of reconciliation required. This is handled by
a few methods inside the Game object, plus a couple of classes in
sim/pitch.py.

Also includes a script to add PitchFx data to game records in the
Stengel database.