Add class for tracking pitcher state
Adding a bare-bones class to track the state of a pitcher (number of
pitches thrown in a game, days since last game, and so forth). I plan
to extend this later with more events, and to add parallel classes for
batters and fielders, but this is enough to get started.