Change pitcher_data values to 2D arrays
The values of the pitcher_data dictionary in PitchDataGenerator were
supposed to be 2D NumPy arrays -- both in my head and according to the
docstrings. However, the actual values were lists of 1D arrays. I fixed
the problem by converting the values to 2D arrays at the end of the
data generation process.