Add install and clean targets to makefile
I've added install and clean targets to the makefile, so I can
eventually go back and get rid of all the ugly path manipulation in the
script. I also need to update the makefile at some point so install and
clean work on more than just Linux for Python 2.7. In the meantime, the
new makefile targets are documented in the README.