Add notebook on Pitchfx variable selection
I'm adding a Jupyter notebook documenting my train of thought for
determining which variables to use in modeling pitch type.

This is the first file in the notebooks/ directory. This directory is
for saving Jupyter notebooks that document some of the statistical
decisions that I make over the course of the project.