Fixed errors in PitchDataGenerator output
I just caught and fixed two major issues/errors in the data output from
the PitchDataGenerator class. First, the order in which elements were
being concatenated was wrong, so columns of data were effectively being
mislabeled. I corrected the concatenation order, so the variable_names
class attribute is the same as before, but is now correct.

The other problem was that no pitches were getting counted as pickoffs.
I changed the test for pickoffs in the Game class to use the
destination instead of the pitch type, and pickoffs are now being
counted appropriately.