Handle edge cases in Player
The initial version of Player had basic functionality, but it wasn't
able to cope with some edge cases. In particular:

The constructor would fail when passed any arguments it wasn't
expecting, which is a problem when constructing from a dictionary in
MongoDB, where there is an _id attribute. Added **kwargs to the
constructor arguments, which is ignored.

add_details_from_retrosheet_page now checks to see if the detail file
exists. If not, it returns False.

add_details_from_retrosheet_page now adds information about the date of
a player's final game in the majors, if available.

add_details_from_retrosheet_page can now handle several unusual cases
in detail files. Birth date, debut, and physical lines are now
determined by matching the start of each line in the table, instead of
assuming fixed indices. Height is sometimes given in half-inches, so
height is now a float instead of an int.

Test cases have been added for all of these scenarios.