Fix rare Retrosheet parsing error
Sometimes, the way to interpret an individual Retrosheet is ambiguous,
and the only way to determine the correct interpretation is to try all
possible options and see which one has the game end at the right time.
It turns out that rare false positives are possible when the simulation
gets one out ahead of reality, then a double play at the end of an
inning pushes the simulation count of outs to 4. Thus, the game ends at
the right time, but one or more half-innings in the middle of the game
are misaligned.

I caught this when I saw some pitches with a current pitcher of None
according to the game status, which should never happen. The fix is to
add logic to check for excess outs in a half-inning, and to include
a check for excess outs in Game.verify_ending().