Add variables from PlayerInfo to PitchData output
It's reasonable to expect that batting and throwing handedness (which
are stored in PlayerInfo objects) would have a significant impact on
the outcome of a pitch. It's easy enough to add all of the other
variables in PlayerInfo at the same time, so I went ahead and added
them, even though their marginal impact will likely be less.