PitchDataGenerator returns array instead of dict
PitchDataGenerator originally returned a dict of arrays, where each
dict contained data for a single pitcher. This made sense when my focus
was on clustering and density estimation by pitcher, but makes less
sense for training a pitch outcome model.

It's also easier (and much less RAM-intensive) to split out single
pitchers' data from the combined array on an as-needed basis than to
combine all of the individual pitchers' data. Thus, this is the
preferred data format going forward.