Refactor PitchDataGenerator
The data for pitch outcome modeling has gotten more and more complex,
and it makes sense to represent the pitch data as separate arrays for
each type of input instead of keeping everything in one giant array.
Additionally, some operations on pitch data, like subsetting and
batching, take a fair bit of work. In light of this, I've created a new
PitchData class to represent the pitch data so it can manage itself.
PitchDataGenerator still exists to create the pitch data from a
database.