Add get_batch method to PitchData
Added get_batch and reset_batch_index methods to PitchData object to
handle the bookkeeping associated with generating batches of data to
feed into a Tensorflow model. get_batch retains state across calls, so
it should iterate through (almost) all of the data in PitchData. If
the number of observations in PitchData is not 0 mod batch_size, the
remainder will never be fed in.