Add as_dict and from_dict methods to PitchData
Pickling a PitchData object in its entirety is convenient, but it has
major drawback of requiring the data to be re-created and re-pickled
any time new behavior is added to the PitchData class. By pickling a
dictionary of the data and unpickling it into a class constructor, we
have the ability to update class behavior without regenerating data.