Add model of pitch outcome
Added a starter model for pitch outcome. The current model is just a
multinomial logistic regression on the numeric inputs, but it serves as
a useful baseline to compare with future modeling efforts.

Model performance is perplexity of ~4.02 on validation data with a
month's worth of training data, versus a null perplexity of ~5.55 and a
naive perplexity of ~4.51.