Add kernel density estimate of pitch distributions
Added a class to fit kernel density estimates of pitch distributions
(on the basis of three dimensions of PitchFx data). Also contains
methods for rendering the estimated density as a 3-dimensional grid,
which may prove useful as an input to the pitch outcome model.

Included a Jupyter notebook documenting selection of kernel bandwidth.