Add pitch density inputs to PitchOutcomeModel
Added pitcher-level estimates of pitch distribution as optional inputs
to the PitchOutcomeModel. Originally, the plan was to feed
three-dimensional grid renderings of the estimated density function to
the model and apply convolutional layers to generate inputs for the
model. However, this plan turned out in practice to be painfully,
unworkably slow.

I compensated by turning the density into a 1-D vector
and feeding it in like any other covariate -- at this point, it's
essentially an extra pitcher-level embedding vector that's not being
estimated simultaneously with the rest of the model.

Runtime with this implementation strategy is acceptable; it's currently
unclear whether including it in the model offers significantly better
performance.

Also reduced resolution of default kernel density render from 30x30x30
to 10x10x10 to cut down on runtime.