Add alternate pitch density representations
The current default pitch density estimate representation is a 10x10x10
array of values. It turns out that the majority of the cells in these
arrays are essentially zero for every pitcher. Since we're not
currently exploiting the spatial nature of the density estimate in
PitchOutcomeModel, we can speed up model fitting by dropping all of the
junk voxels.

Another, even sparser representation of the pitch density is to simply
provide key quantiles for each of the three dimensions we're
considering, and completely ignore the joint distribution.

Added cells to the density estimation notebook to generate both of
these representations.