Add save and restore methods to PitchOutcomeModel
Added save and restore methods to PitchOutcomeModel, so the state of a
trained model can be saved, instead of needing to be re-fit the model
object is recreated.

Also significantly refactored the graph and session fitting code. The
default graph for self.session is now self.graph, as I had intended,
eliminating the with self.graph.as_default() context manager in the
train method.

Also moved the variable initialization logic from train to the
constructor, where it belongs.