Made dropout a parameter in PitchOutcomeModel
The original implementation of dropout in PitchOutcomeModel was crude
for two reasons. First, it forced the dropout value to be 0.5, instead
of allowing the user to specify it. Second, and more importantly, the
dropout rate was a fixed parameter, meaning that dropout was happening
when the model was being scored on validation or test data. This should
not be happening -- dropout is purely a training operation!

Fixed both of these problems by creating a constructor argument for
dropout_rate, adding a placeholder in the graph for the dropout rate,
and forcing keep_prob=1 on the score() and predict() methods.