Fix serialize bug in Python 3
The test suite was failing when I migrated to Python 2 to 3, and it
turns out it was because I was deleted keys from a dictionary while
iterating through it. Apparently, this is totally kosher in 2, but will
throw an error in 3.

Fixed the problem by creating a list of keys to delete inside the loop
and deleting all of them afterwards.