Fix Python 3 error in Bases
There was a line in Bases that caused an error in Python 3, but which
was fine in Python 2. I didn't catch it until now because this portion
of the codebase isn't covered by unit tests yet.

As originally written, the code relied on a <= comparison between two
values, one of which could be None. Apparently, this is not allowed in
Python 3. The method was written to avoid that comparison if the value
was, in fact, None.