Stop multiple updates in add_pitchfx
The add_pitchfx script was throwing an error when the progress bar was
being passed a value that was greater than the specified maximum. This
was unexpected behavior, because the loop that the progressbar was
tracking only updated records in the collection, and I verified that
the cardinality of the collection remained constant during the loop.

As it turns out, some records were getting updated twice -- for example,
HOU200904070, STL200904070, SEA200904070, then HOU200904070. This was
completely unexpected behavior for me, and I'm still not entirely sure
why it was happening, but here's my best guess. When MongoDB is pulling
records, from a set, it doesn't keep track of which records it's
pulled -- it instead relies on some external cue, like the order of the
records in the filesystem, and iterates through in that order. If we're
adding a lot of information to the game record and increasing its size
to the point that MongoDB has to reallocate space for it during the
update operation, then the "new" record might wind up at the end of the
list, and MongoDB would happily retrieve it again, not knowing it had
already retrieved it.

Regardless of the true root cause, the fix is pretty simple. I added a
list of game IDs that have been processsed. If a retrieved game ID has
already been processed, we continue. Note that we can't rely on
something like events[0].pitch_fx not being None -- first off, we may
want to intentionally overwrite pre-existing PitchFx data, and second,
the first pitch may be missing PitchFx data.

I know that the runtime of my solution is something like O(n^2), but
at this point, it's far from the limiting factor, and on the size of
data we'll ever be working with -- tens of thousands of games -- it
shouldn't ever be a serious performance bottleneck.

While I was at it, I also removed the update() method from game and
but the corresponding logic directly in the main loop of add_pitchfx.
Making update() a method of game had been bothering me for a while,
and this was a convenient opportunity to fix it.