Fix pitch data destination path creation
This script had a bug in it at the creation of destination_directory.
os.path.join was getting passed a string instead of a list, so it
returned a path to a deeply nested set of one-character directories.
This also caused problems when the (real) destination directory didn't
exist, because the following call to open() would fail.

Also removed the non-Python-3 compliant use of file() and replaced it
with open().