Add game dates to PitchData
Added an extra attribute to PitchData with the date of the game that
each pitch was thrown in. This information isn't part of the pitch_data
array because it will never be used for predictive purposes, but it is
extremely useful for filtering purposes.

The game dates are being stored as a NumPy array of type "datetime64".
According to the NumPy documentation, this feature is still
experimental and subject to change. Here's hoping this won't be the
cause of future instability, because it makes my code a lot cleaner
right now.